Travis: explicitly specify flags in cabal install
After much head scratching, it turns out that `cabal install` doesn't
honour flag defaults consistently. While `mond` (and possibly `metad`)
flags are honoured (PSQueue and snap-server are installed), the
`htest` flag is ignored and thus the test-framework dependencies are
not installed.

I've tested this with a most recent Debian sid, and I presume the same
happens with the "trusty" environment in Travis. So let's explicitly
enable all the important flags.

Signed-off-by: Iustin Pop <iustin@k1024.org>