CabalDependencies: move to getPersistBuildConfig
The use of "tryGetPersistBuildConfig" would make sense if we can
recover from failure, but in this case, the only thing it does is to
hide the actual error and replace it with a generic and not-so-helpful
message. I think it's much better to use the plain 'get' version,
which in case of failure will throw an IO error that will also
terminate the program, but with a detailed error message.

Example new failure (before it was just "cannot read
dist/setup-config"):

    CabalDependenciesMacros.hs: Saved package config file is outdated:
    • the Cabal version changed from Cabal-2.0.1.1 to Cabal-2.2.0.1
    • the compiler changed from ghc-8.2 to ghc-8.4
    Re-run the 'configure' command.

Signed-off-by: Iustin Pop <iustin@k1024.org>