Fix shell tests with shelltestrunner 1.9
shelltestrunner 1.9 introduced 2 new test formats which are tried before
the pre-1.9 format we are using. One of these formats features "short"
delimiters, with $ being one of them. All our test commands begin
with $ (referencing an environment variable), which causes
shelltestrunner 1.9 to assume the test file is in version 2 and
subsequently break because we're using "long" delimiters:

 :parse error in /tmp/gntrelease.n5aNp3p7SZ/dist/ganeti-2.16.1/_build/sub/../../test/py/ganeti-cli.test: [Failed]
 "/tmp/gntrelease.n5aNp3p7SZ/dist/ganeti-2.16.1/_build/sub/../../test/py/ganeti-cli.test" (line 3, column 2):
 unexpected ">"

I'm not sure whether environment variables are intentionally supported,
or just worked sofar as a byproduct of passing the command line through
a shell. In any case, fix this by using `sh -c` to run the test
commands. This is admitedly not pretty, but allows us to retain
compatibility with previous shelltestrunner versions.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>