Replace tuple parameter unpacking instances
Tuple parameter unpacking, as supported in Python 2, allows the
specification of composite function arguments, such as:

 def f(x, (y, z), a):

where (y, z) is expected to be a single argument consisting of a tuple
of length two, which is decomposed to y and z for all subsequent code.

Python 3 has removed support for tuple parameter unpacking with
PEP-3113. The proposed migration for such code is to replace the above
definition with something like:

 def f(x, y_z, a):
   (y, z) = y_z

Replace all instances of tuple parameter unpacking to allow our existing
code to work with Python 3. Note that this is a best-effort solution:
arguably we should move away from tuple arguments completely, as they
are notorious for introducing bugs due to e.g. using the wrong element
order.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>