Convert utils.io.WriteFile to work with bytes
WriteFile would accept either unicode, or str and write it out to the
file specified using low-level APIs (os.write). In the Python 3 world,
the function must handle str and bytes instead, which makes things a bit
more complicated: while in Python 2 only unicode objects had to be
encoded (which was rare), in Python 3 we have to encode *all* strings to
bytes, and in doing so, we must avoid memory explosion; a simple `data =
data.encode()` will not do if `data` is big enough.

Refactor the write loop to utilize Python's buffered I/O system (via
open()), as it allows us to do on-the-fly encoding of str to bytes when
opening a file in text mode. Additionally it makes the write loop
simpler.

Also, to make Ganeti more robust, set the encoding for writing files to
UTF-8, independent of what the system locale might be.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>