Drop ConfigObject.__eq__
TL;DR: we can't have *both* a __hash__ *and* a "real"¹ __eq__ defined
for mutable objects like ConfigObject. We're depending on __hash__ for
inclusion of ConfigObjects in sets, so we need to drop __eq__, without
breaking the tests that rely on it.

¹as in "relying on actual object contents and not on id()"

Citing the official Python docs[1]:

  If a class does not define an __eq__() method it should not define a
  __hash__() operation either; if it defines __eq__() but not __hash__(),
  its instances will not be usable as items in hashable collections. If a
  class defines mutable objects and implements an __eq__() method, it
  should not implement __hash__(), since the implementation of hashable
  collections requires that a key’s hash value is immutable (if the
  object’s hash value changes, it will be in the wrong hash bucket).

  User-defined classes have __eq__() and __hash__() methods by default;
  with them, all objects compare unequal (except with themselves) and
  x.__hash__() returns an appropriate value such that x == y implies both
  that x is y and hash(x) == hash(y).

  A class that overrides __eq__() and does not define __hash__() will have
  its __hash__() implicitly set to None. When the __hash__() method of a
  class is None, instances of the class will raise an appropriate
  TypeError when a program attempts to retrieve their hash value, and will
  also be correctly identified as unhashable when checking isinstance(obj,
  collections.abc.Hashable).

According to the last paragraph, in Python 3 our ConfigObjects do not
have a __hash__ method because ConfigObject defines __eq__ but not
__hash__. __eq__ was introduced in 19830e88 to facilitate argument
comparison in assert_called_once_with() mock calls. On the other hand,
we rely on __hash__ whenever we use ConfigObjects in sets and set-like
structures (which is often enough), so the reasonable solution would be
to work around the lack of __eq__ rather than live with a lack of
__hash__.

Since __eq__ is only there to facilitate mocks, be explicit about
matching arguments in tests and introduce a ConfigObjectMatcher class
that wraps a ConfigObject to provide __eq__ and move __eq__ to the
ConfigObjectMatcher class.

[1] https://docs.python.org/3/reference/datamodel.html#object.__hash__

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>