hs-tests: generate network addresses without host bits
Python 2's ipaddr was pretty liberal when it came to networks: it would
happily accept network addresses with host bits set, e.g. 192.0.2.1/24.
This is no longer the case with Python 3's ipaddress, which complains by
default when a network address has host bits set:

  >>> import ipaddress
  >>> ipaddress.IPv4Network('192.0.2.1/24')
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "/usr/lib/python3.7/ipaddress.py", line 1528, in __init__
      raise ValueError('%s has host bits set' % self)
  ValueError: 192.0.2.1/24 has host bits set

While there is a kwarg (strict=False) to disable this check, I think
it's a good idea to respect this constraint; we just need to be careful
when generating random network addresses.

This commit fixes IPv4 and IPv6 network generators to always return
network addresses with the host bits cleared.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>