Port Ganeti.Hash to cryptonite
There's a bug exposed by the combination of Crypto and GHC 8.6 that
results in an endless loop[1], caused by GHC 8.6 changing accumArray
from lazy to strict. The endless loop can be reproduced as follows:

 Prelude> :m +Data.HMAC
 Prelude Data.HMAC> hmac_sha1 [] []

Unfortunately Crypto was last updated in 2012 and seems to be largely
unmaintained, so at this point it's probably a good idea to switch our
HMAC implementation over to cryptonite, which is the de facto standard
for crypto in Haskell these days.

[1] https://gitlab.haskell.org/ghc/ghc/issues/16406

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>