QuickCheck 2.12 compatibility
QuickCheck 2.12 has some breaking changes, in particular it swapped the
order of the two first arguments of `cover' and changed the percentage
from Int to Double.

For the first change, adapt the code to the new argument order and
provide a compatibility function for older QuickCheck versions.

For the latter, reflect the type change in our signature of stableCover
and handle the conversion to Int for compatibility with older versions
in cover'.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>