Skip a couple of filestorage tests if run as root
A couple of filestorage tests, namely
testCheckFileStoragePathNotWritable and testFileCreation attempt to
check for permission-denied conditions, which cannot be triggered when
run as root. Do a best-effort attempt to skip running these tests when
or current UID is zero. Note that the tests should normally not be run
as root, however there are some container implementations (e.g. GitHub
actions) that run things as root inside the containers by default.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>