Avoid using forkProcess for job process creation
forkProcess is currently a big can of worms. It's widely known to be a
complete no-go when using a threaded runtime; furthermore, judging from
the comments in src/Ganeti/Query/Exec.hs, it needs a lot of workarounds
(like retrying fork()!) to be made reliable. But most importantly, it
seems to interact badly with I/O done around a fork, to the extent that
it is extremely fragile and can cause ill side-effects (see issues #1323
and #1410).

Moving away from forkProcess means relying on
System.Process.createProcess and friends to do subprocess management.
This is a stable interface, implemented in C, that is known to work
well with both, threaded and non-threaded runtimes. However, this is an
atomic fork()/exec() implementation, which means that we can directly
execute the Python interpreter, but we're not going to be able to
perform any child process setup on the Haskell side.

At least to my eyes, all of the setup steps performed in runJobProcess
can be performed in Python with equal ease/reliability, and I see no
particular reason why livelock creation must be done in Haskell.
Afterall, livelock creation is still performed in the child process and
has to be communicated back to the master via a pipe, so there's nothing
there that can't be done with Python.

So, let's get rid of the whole forkProcess dance by invoking the Python
interpreter directly using System.Process.createProcess. This requires a
slight change to the master <-> job protocol: instead of learning the
livelock name from the master, the job process creates the livelock
itself and sends the name back to the master in order to have it written
to the jobfile. Also, there's no reason to get the job ID from the
master, since we already have it in argv[].

This change makes the code in src/Ganeti/Query/Exec.hs considerably
cleaner, and completely eliminates the sporadic 30-second pauses
observed in #1410. A simple test with 500 jobs reveals a stable queueing
delay of 0.2-0.3 seconds, without 30-second spikes.

 find /var/lib/ganeti/queue -type f -mmin -5 | \
	xargs cat | \
	jq --raw-output '((.start_timestamp[0] * 1000000 + .start_timestamp[1] \
	  - .received_timestamp[0] * 1000000 - .received_timestamp[1]) / 100000) | \
	floor | ./10' | sort -g | uniq -c
    278 0.2
    222 0.3

Reproducing the double writes of #1323 is much harder, but I have
reasons to suspect that they will be fixed as well.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>