tests: factor out b64encode_string
There are various places in the tests where we want to replace a string
with its base64-encoded equivalent. Under Python 3, this requires
converting to/from bytes, which makes the code a bit ugly. Add a helper
function, testutils.b64encode_string, that accepts and returns strings,
and adjust the code where necessary.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>