getNodeInstances: count each instance at most once
Commit 9825767a ("Tune getNodeInstances DRBD secondary computation")
refactored the way we compute secondary instances for a node, by
iterating over an instance's disk and extracting the secondary node.
Unfortunately, this introduced a bug where each instance is counted as
many times as its disk count.

We want to consider each instance only once, so just check if any of the
instance disks have the node in question as secondary.

This fixes #1399.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>