gnt_job: do not encode job summary
gnt-job does a best-effort attempt to handle unicode by calling
str.encode(). In Python 2 this did the right thing, but with Python 3
it converts the string to bytes, which is then output as a bytes
literal. Since Python 3 works with unicode strings internally, there's
no need to encode anything anymore, so adjust accordingly.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>