import-export: use binary mode for unbuffered I/O
import-export uses os.fdopen to explicitly create unbuffered file
handles. However, this is only possible for handles created in binary
mode, as encoding to/from text possibly needs to reassemble multi-byte
sequences. To this end, Python 3 throws an exception when os.fdopen is
called with a text mode and buffer size set to zero.

Fix this by making sure we use "rb" for unbuffered handles.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>