utils.text: make LineSplitter handle bytes
LineSplitter is used by the import/export daemon to reassemble program
output in whole lines. As outlined in a previous commit ("import-export:
use binary mode for unbuffered I/O"), I/O from processes is performed
using binary streams rather than text. Make LineSplitter handle
bytestreams up to the point that it can split them into lines.

Note that we cannot simply decode() data as it arrives, because a
multi-byte character might be split across two chunks. However, we can
safely decode data after we have split the input in lines, as we can
assume that multibyte characts will not be split across lines (and if
they are, the input is malformed anyway). Also note that because we are
decoding program output, we let Python decide which encoding to use,
based on the system locale.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>