the py-tests target now evaluates if tests are possible
The Haskell test target evaluates the HS_NODEV variable which defines if
all required modules for testing are available. `configure` defines the
same for Python tests but until now it was never evaluated by `make`.
With this change, people will run into a more descriptive error if one
of the required Python modules is missing.

Signed-off-by: Rudolph Bott <r@spam.wtf>