Incorrect evaluation of PY_NODEV in `configure`
The PY_UNIT variable defines if Python unit tests are possible. This is
evaluated from the variable PY_NODEV which is either empty (good) or
contains a list of missing Python modules (bad). This patch flips the
logic so PY_UNIT will be set to `true` on an empty list.

Signed-off-by: Rudolph Bott <r@spam.wtf>