Fix VPATH builds using tarballed sources
Commit edc49e6a4 enabled AM_MAINTAINER_MODE to allow out-of-source VPATH
builds with a read-only source directory. This works fine when the
original source directory is a git repository, but fails when a tarball
source without .git/ is used due to lack of vcs-version in the build
directory.

vcs-version is generated from git-describe and included in the tarballs,
but never copied to the build directory in VPATH builds. To make VPATH
builds work, use a full path including $(srcdir) when accessing
vcs-version.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>