check-imports: freeze standard module list
check-imports attempts to ensure that certain stand-alone programs do
not import ganeti in their code. It does so as follows:

 1. It builds a list of "standard" Python modules as soon as it starts.
 2. It loads some ganeti code
 3. For each of the programs, it goes through the module list, removes
    all non-standard modules, loads the program, and checks that
    "ganeti" appears nowhere in the module list.

Step 1 is implemented using sys.module.keys(). In Python 3, dict.keys()
returns a dictionary view object[1], which provides a _dynamic_ view of
the dictionary data. In contrast, Python 2's dict.keys() would return a
simple, static list of the dictionary keys at the time of invocation.

This means that in Python 3 the list of standard modules will always
match the list of current modules (since it's a dynamic view on
sys.modules), so no modules will be unloaded at step 3, and the run will
fail with an assertion error because "ganeti" will still be loaded.

Fix this by persisting a snapshot of the standard module list in a set().

[1] https://docs.python.org/3/library/stdtypes.html#dictionary-view-objects

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>