Fix Makefile/dircheck to ignore hidden folders
check-local checks if all folders are known to the Makefile. It has a
list of exceptions (DIRCHECK_EXCLUDE) which does not support subfolders.
Hence, the .git folder already had a special treatment in the related
`find` statement. This commit generalizes this to ignore all hidden
folders (e.g. .git, .github, .vscode etc.)

Signed-off-by: Rudolph Bott <r@spam.wtf>