support Haskell process before 1.6.3
Commit 219b2ab9 moves away from unreliable forkProcess to
System.Process.createProcess and friends. The process library
>=1.6.3 introduced the getPid function
(https://hackage.haskell.org/package/process-1.6.3.0/changelog),
which is used in the mentioned commit. On Ubuntu 18.04 there is only
process-1.4.3 available. This commit introduces a wrapper function,
that conditionally switches to an included implementation, if
process is < 1.6.3. This will support Ganeti-3.0 builds from source
on Ubuntu Bionic.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>