Use `cabal v1-configure' when supported
Cabal 3.x changed the behavior of almost all core command to the
so-called "new-style" nix build commands. The old commands are still
kept around as legacy commands prefixed with "v1-". Our build system
currently requires reading in dist/setup-config which is only produced
by the v1 configure command, so add a check to use the v1-configure
command when supported.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>