KVM: improved migration caps handling
Currently we set KVM migration capabilities only on the source instance
and leave them set even if migration fails. This leads to two problems:

 - There are a number of features (e.g. postcopy-ram) where the
   capability needs to be set on both, source and destination instance

 - If a migration breaks because of a capability mismatch, the
   capability remains set causing all future migration attempts to fail
   as well (see Debian bug #969028 [1])

Fix both issues by setting all capabilities on both sides of the
migration and clearing all capabilities on the surviving side when the
migration is over.

[1] https://bugs.debian.org/969028

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>