KVM: fix multi-queue tap creation
We create multi-queue tap interfaces in a loop, by passing the tap name
obtained by the first TUNSETIFF ioctl() to the next ioctl() calls.
This loop assumes that the `name` variable is a string, but goes on to
reassign bytes to it causing the next iteration to fail with an
AttributeError.

Fix this by using a separate bytes variable to hold the interface name
to/from the ioctl() request.

This issue was reported by magnus1 on freenode.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>