KVM: use qemu monitor to set spice password instead of QMP
Before this patch ganeti used the more modern QEMU Machine protocol to set
the spice password for an instance upon startup. This is in contrast to the
method used when setting a VNC password, which employs qemu monitor to
the same effect.

However, it appears that this implementation is currently broken. Instances
with a set spice password will fail to start due to QMP timeout.

This patch changes the method used to set a spice password for the instance
to match the one used for VNC, enabling instances with password-protected
spice servers to start without manual user intervention.

Signed-off-by: George Diamantopoulos <gdiamantopoulos@modulus.gr>