automatic postcopy-ram live migration
This implements automatic postcopy-ram live migration for the KVM
hypervisor. Basically it is a cherry-pick from commit 041ed56d9 adopted
to today's code. Credit goes to the original author Calum Calder.

It extends the migration status by dirty_sync_count, which is the number
of times that dirty ram was synchronized (since qemu-2.1). If the
migration capability postcopy-ram is set and dirty_sync_count >=2, live
migration automatically switches to postcopy-ram migration.

Co-authored-by: Andre Liebigt <andre@cilq.net>
Signed-off-by: Sascha Lucas <sascha_lucas@web.de>