cfgupgrade: inherit target version from autotools
There's no reason to explicitly set the target version in cfgupgrade
every time we do a version bump. Instead, use the constants generated by
autotools.

While at it, also set the downgrade version to the previous (major,
minor) combination. This will work fine for all of the 3.x lifecycle
from 3.1 and on.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>