KVM: fix disk hotplugging with QEMU 4.0+
As of QEMU 4.0, file-backed block devices are opened in read-only mode
and upgraded to read-write mode the first time a write needs to happen.
QEMU calls this "dynamic auto-read-only" and it can be turned off by
appending `auto-read-only=off` to a drive definition.

For auto-read-only to work with fdsets, the fdset backing a drive must
now include two file decriptors, one in read-only and one in read-write
mode, but Ganeti only passes a read-write file descriptor.

We could disable auto-read-only for hotplugged devices, or we can pass a
second, read-only fd in the same fdset. The former would require
version-specific overrides once more and would go against QEMU defaults,
so let's do the latter.

Since QEMU only accepts individual descriptors in `add-fd`, we need to
slightly change _AddFd to allow using the fdset allocated by QEMU for
both descriptors.

This fixes #1547.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>