KVM: Fix regexps for vhost and virtio queues tests
In newer qemu versions the `--help` output regarding `vhost=` and `fds=`
switched from the `-net` to `-netdev tap`. The old option `-net` just
omits a deprecation warning:

 old way to initialize a host network interface
 (use the -netdev option if possible instead)

This patch fixes the regex to use `-netdev tap` instead. Additionally
the test for `vhost=` is fixed, which always matched in the past by
accident (matching an alternative `off` without escaping the `|`).

The regex was successfully tested with qemu-2.11 to 4.2 (Bionic,
Focal).