add rename function to networks
Currently it is not possible to rename a network, besides removing and
recreate it with the new name. In order to remove a network, it must not
contain instances and is not connected to any node group. During
removing/recreating the network information like network address/mask,
gateway, link, vlan etc. must be saved externally. With this change
networks can be renamed like most other Ganeti config objects.