Assessment of cluster-level hypervisor parameters during cluster-verify
This commit implements the assessment of hypervisor parameters set on
the cluster-level. It is achieved through a new method of the hypervisor
abstraction which gets called for each enabled hypervisor during the
`gnt-cluster verify` run. Currently it is only implemented for the KVM
hypervisor and tries to find suboptimal values for e.g. `cpu_type`.

Findings are printed in a new section of the `verify` output and are
only informational, they do not have any impact on the return code of
the command (and can be turned off with a new option).

This feature is meant to uncover parameter values which might have
become deprecated over time but have been preserved during cluster
upgrades. It may also help new users to uncover potential 'bad'
configuration, such as running instances with the 'qemu64' CPU type.

Signed-off-by: Rudolph Bott <r@spam.wtf>