relax VLAN check with VETHs
Normally one can not use identical VLAN on VLAN aware bridge and
traditional VLAN interfaces connected to the same lower device. A common
workaround is to insert a veth pair between the VLAN aware bridge and
traditional VLAN interfaces. This way one can stack VLAN aware bridge
and traditional bridges, too. This commit relaxes the VLAN check, to be
skipped, if the lower interface is of type `veth`.