KVMs HotplugSupported() checks QEMU version via QMP without connecting first
HotplugSupported tries to retrieve the `version` attribute of the
QMPConnection object before it is assigned as there is no active
connection yet. In all other cases the QMP socket is accessed through
QMPconnection methods which use the `@_ensure_connection` decorator.
This commit introduces a GetVersion() method for QMPconnection which
also uses the same decorator to work around the issue of an
uninitialiased `version` attribute.

Signed-off-by: Rudolph Bott <r@spam.wtf>