Allow to run QA suite's instance tests with alternating HV parameters
The QA suite does not test different hypervisor parameters (e.g.
`use_chroot`, `use_guest_agent` etc.). This commit implements the
automatic cycling through values of boolean parameters and parameters
which values are known to Ganeti (e.g. `disk_cache`).

It does not cover parameters which accept custom values and it also
does not cover parameters which need to be set along with others (e.g.
`security_model` and `security_domain`). However, this could be
implemented on top of this patch later on.

Signed-off-by: Rudolph Bott <r@spam.wtf>