Warn users that changing DRBD parameters does not affect running instances
When a user changes DRBD parameters (e.g. sync-related parameters) they
will be only used in future DRBD setups and not applied to currently
existing ones. This only adds a warning. A better solution would be to
have `activate-disks` apply changed parameters where possible and either
add this information to the warning message or remove it and run
activate-disks automatically for all existing devices.

Partially fixes #781

Signed-off-by: Rudolph Bott <r@spam.wtf>