py-tests: make tests compatible with roman 3.2+
The roman library (used to format numbers to roman numerals) has changed
its behaviour starting with version 3.2 and now formats a single digit
`0` as 'N' instead of `0`. This commit broadens the test to accept both
values. As roman numerals are purely used in output and are never parsed
as input, this should be fine.

Signed-off-by: Rudolph Bott <r@spam.wtf>