KVM: make StopInstance() non-blocking again
This fixes a misunderstanding introduced in 536df7481/#1662 - waiting
for the instance to exit cleanly should not be implemented in the
hypervisor code because it is already present one layer above in the
backend method InstanceShutdown(). Said method is also responsible for
re-requesting shutdown if the instance has missed/failed to react to
ACPI power button events.

Signed-off-by: Rudolph Bott <r@spam.wtf>