HTTP/TLS implementation: fix OpenSSL exceptions during portprobing
When a client connects and disconnects without any data transmitted the
internal SSL/TLS server implementation throws a bunch of exceptions. This
commit catches the specific 'SSL shutdown while in init' condition, and
adds a warning message to the log.

This specifically addresses the probing of noded reachability between
all nodes during `gnt-cluster verify` which generates many execeptions
in the noded log, depending on the cluster size.

Signed-off-by: Rudolph Bott <r@spam.wtf>