QA: fix SSH multiplexing setup
SSH multiplexing was broken in the QA suite, becuase it only used to set
up a multiplexing connection for the first node, node for all nodes.
Additionally, the QA actually has two instances of the qa_utils module
loaded (due to importing the module in general and also importing select
functions like AssertCommand directly). This lead to the situation that
the dictionary qa_utils._MULTIPLEXERS is either populated or empty,
depending on the callstack. By moving it to the qa_config module this
situation is avoided.

Signed-off-by: Rudolph Bott <r@spam.wtf>