uidpool_unittest: avoid using negative UIDs
The uidpool unittests long relied on the fact that pgrep would accept
"-1" as a valid uid and in turn would match nothing. As of propcs 4.0.3
however, pgrep will error out when passed a negative uid, thus breaking
the tests.

Work around this by using a valid UID, but one that is unlikely to be
used in real systems: on most systems, SUB_UID_MAX which is the highest
uid expected to be in regular use, is set to 600100000. Picking 2^30 +
42 looks like a reasonably safe choice.

This fixes #1691.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>
(cherry picked from commit 9cd67e6a81c66ed326d68ea8c3241d14eea6550b)