daemon: drop event scheduler
The main loop event scheduler was introduced with e2be81cf2 (Implement
timers in confd, 2009-08-23) to support timed events for ganeti-confd
(at the time a Python daemon). Since the conversion of confd to Haskell,
this asyncore-specific code has been unused and it makes sense to remove
it as asyncore will be dropped in Python 3.12. Drop all dead code
together with the relevant tests.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>