impexpd: verify remote socket against actual host name
In 7bb0351 (impexpd: fix certificate verification with new socat
versions, 2017-12-20), a hostname verification was introduced to fix
socat's new (and proper) behavior of actually checking the remote
hostname during OpenSSL-protected transfers.

The problem, however, is that the hostname used was the default
`X509_CERT_CN` constant (`x509CertCn` in Haskell) which is hardcoded
to `ganeti.example.com`. In a real-world deployment, it seems like the
remote CommonName (CN) of the certificate used by the export daemon is
actually the target node name.

In my case, it meant I was getting the following error from socat
during transfers:

    Disk 0 failed to send data: Exited with status 1 (recent output: socat: E certificate is valid but its commonName does not match hostname "ganeti.example.com")

At first I thought socat might be doing us some trouble, but no: socat
works properly. An example is this:

 1. generate a self-signed certificate:

        openssl req -x509 -newkey rsa:4096 -keyout key.pem -out cert.pem -sha256 -nodes -subj '/CN=ganeti.example.com' -days 1

 2. start a socat server:

        socat -ls -d -d  STDIO OPENSSL-LISTEN:8443,reuseaddr,forever,key=key.pem,cert=cert.pem,cafile=cert.pem

 3. connect to it with a client:

        socat -ls -d -d   STDIO  OPENSSL:localhost:8443,openssl-commonname=ganeti.example.com,verify=1,key=key.pem,cert=cert.pem,cafile=cert.pem

... which actually works, which means the `openssl-commonname`
argument actually works, and works properly. If it's changed, for
example, to `ganeti.example.net`, the above fails with the
aforementioned error message.

We fix this by doing a reverse name resolution on the provided IP
address. Now, we don't *assume* it's an IP address: this code kicks in
only if the impexpd is passed an actual IP address, but in my
experience it seems to always be the case (which is probably a
separate problem to fix).

This is rather brittle and assumes DNS will not lie, which is quite a
stretch. In our environment, however, we have end-to-end DNSSEC so we
can trust the DNS. And this beats hardcoding verify=0, which is the
other workaround that can be done to fix this issue.

Closes: #1681