lib/runtime: improve not-master error
For unclear reasons, there was an environment where the master was
running, but the clients did not believe it was the master.

kernel.hostname returned the unqualified hostname, as expected, but the
master check was looking for the FQDN.

Improve the error message as a start on making it easier to debug.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>